Require task queue to be attached to completes/heartbeats
Dramatically simplify activity task tracking
