+++ src/worker/activities.rs
@@ -4,13 +4,14 @@ use crate::{
     pollers::BoxedActPoller,
     protos::{
         coresdk::{
+            activity_result::{self as ar, activity_result},
             activity_task::{ActivityCancelReason, ActivityTask},
             ActivityHeartbeat,
         },
         temporal::api::workflowservice::v1::PollActivityTaskQueueResponse,
     },
     task_token::TaskToken,
-    ActivityHeartbeatError, PollActivityError, ServerGatewayApis,
+    ActivityHeartbeatError, CompleteActivityError, PollActivityError, ServerGatewayApis,
 };
 use activity_heartbeat_manager::ActivityHeartbeatManager;
 use dashmap::DashMap;
@@ -122,14 +123,67 @@ impl WorkerActivityTasks {
         }
     }
 
-    /// Mark the activity associated with the task token as complete, returning the details about it
-    /// if it was present.
-    pub(crate) fn mark_complete(&self, task_token: &TaskToken) {
-        self.heartbeat_manager.evict(task_token.clone());
-        self.activities_semaphore.add_permits(1);
-        self.outstanding_activity_tasks
-            .remove(&task_token)
-            .map(|x| x.1);
+    pub(crate) async fn complete(
+        &self,
+        task_token: TaskToken,
+        status: activity_result::Status,
+        gateway: &(dyn ServerGatewayApis + Send + Sync),
+    ) -> Result<(), CompleteActivityError> {
+        let (res, should_remove) = if let Some(is_known_not_found) = self
+            .outstanding_activity_tasks
+            .get(&task_token)
+            .map(|t| t.known_not_found)
+        {
+            // No need to report activities which we already know the server doesn't care about
+            if !is_known_not_found {
+                let maybe_net_err = match status {
+                    activity_result::Status::Completed(ar::Success { result }) => gateway
+                        .complete_activity_task(task_token.clone(), result.map(Into::into))
+                        .await
+                        .err(),
+                    activity_result::Status::Failed(ar::Failure { failure }) => gateway
+                        .fail_activity_task(task_token.clone(), failure.map(Into::into))
+                        .await
+                        .err(),
+                    activity_result::Status::Canceled(ar::Cancelation { details }) => gateway
+                        .cancel_activity_task(task_token.clone(), details.map(Into::into))
+                        .await
+                        .err(),
+                };
+                match maybe_net_err {
+                    Some(e) if e.code() == tonic::Code::NotFound => {
+                        warn!(task_token = ?task_token, details = ?e, "Activity not found on \
+                        completion. This may happen if the activity has already been cancelled but \
+                        completed anyway.");
+                        (Ok(()), true)
+                    }
+                    Some(err) => (Err(err), false),
+                    None => (Ok(()), true),
+                }
+            } else {
+                (Ok(()), true)
+            }
+        } else {
+            warn!(
+                "Attempted to complete activity task {} but we were not tracking it",
+                &task_token
+            );
+            return Ok(());
+        };
+
+        if should_remove
+            && self
+                .outstanding_activity_tasks
+                .remove(&task_token)
+                .is_some()
+        {
+            self.activities_semaphore.add_permits(1);
+            self.outstanding_activity_tasks
+                .remove(&task_token)
+                .map(|x| x.1);
+            self.heartbeat_manager.evict(task_token);
+        }
+        Ok(res?)
     }
 
     /// Attempt to record an activity heartbeat
