+++ src/worker/activities/activity_heartbeat_manager.rs
@@ -247,6 +247,7 @@ impl ActivityHeartbeatManager {
 #[cfg(test)]
 mod test {
     use super::*;
+    use crate::test_help::TEST_Q;
     use crate::{
         pollers::MockServerGatewayApis,
         protos::{
@@ -347,6 +348,7 @@ mod test {
         match hm.record(
             ActivityHeartbeat {
                 task_token: vec![1, 2, 3],
+                task_queue: TEST_Q.to_string(),
                 details: vec![Payload {
                     // payload doesn't matter in this case, as it shouldn't get sent anyways.
                     ..Default::default()
@@ -372,6 +374,7 @@ mod test {
         hm.record(
             ActivityHeartbeat {
                 task_token,
+                task_queue: TEST_Q.to_string(),
                 details: vec![Payload {
                     metadata: Default::default(),
                     data: vec![payload_data],
