+++ tests/integ_tests/heartbeat_tests.rs
@@ -45,6 +45,7 @@ async fn activity_heartbeat() {
         sleep(Duration::from_millis(100)).await;
         core.record_activity_heartbeat(ActivityHeartbeat {
             task_token: task.task_token.clone(),
+            task_queue: task_q.to_string(),
             details: vec![],
         });
     }
@@ -56,6 +57,7 @@ async fn activity_heartbeat() {
     // Complete activity successfully.
     core.complete_activity_task(ActivityTaskCompletion {
         task_token: task.task_token,
+        task_queue: task_q.to_string(),
         result: Some(ActivityResult::ok(response_payload.clone())),
     })
     .await
