+++ tests/load_tests.rs
@@ -9,7 +9,6 @@ use temporal_sdk_core::{
         ActivityTaskCompletion,
     },
     prototype_rust_sdk::WfContext,
-    tracing_init,
 };
 use test_utils::CoreWfStarter;
 
@@ -17,8 +16,6 @@ const CONCURRENCY: usize = 1000;
 
 #[tokio::test]
 async fn activity_load() {
-    tracing_init();
-
     let mut starter = CoreWfStarter::new("activity_load");
     starter
         .max_wft(CONCURRENCY)
@@ -26,7 +23,7 @@ async fn activity_load() {
         .max_at_polls(10)
         .max_at(CONCURRENCY);
     let worker = starter.worker().await;
-    let task_q = &starter.get_task_queue().to_owned();
+    let task_q = starter.get_task_queue().to_owned();
 
     let activity_id = "act-1";
     let activity_timeout = Duration::from_secs(8);
@@ -78,7 +75,8 @@ async fn activity_load() {
     let all_acts = async move {
         let mut act_complete_futs = vec![];
         for _ in 0..CONCURRENCY {
-            let task = c2.poll_activity_task(task_q).await.unwrap();
+            let task_q = task_q.clone();
+            let task = c2.poll_activity_task(&task_q).await.unwrap();
             assert_matches!(
                 task.variant,
                 Some(act_task::Variant::Start(ref start_activity)) => {
@@ -90,6 +88,7 @@ async fn activity_load() {
             act_complete_futs.push(tokio::spawn(async move {
                 core.complete_activity_task(ActivityTaskCompletion {
                     task_token: task.task_token,
+                    task_queue: task_q,
                     result: Some(ActivityResult::ok(pd.into())),
                 })
                 .await
